Explicitly call Client.postConstruct(Credentials).
The class was previously called by Guice as part
of the dependency injection. This has to be done
manually now. If not, the application will crash
during startup with an IndexOutOfBoundsException.

Reason being that none of the shards are loaded,
so the database can't add the owner to the database.